--- frontend-vue/src/views/contracts/ContractsPage.vue
@@ -1,19 +1,0 @@
-<template>
-    <v-container fluid>
-        <contracts-table />
-    </v-container>
-</template>
-
-<script>
-import { defineAsyncComponent } from 'vue';
-
-export default {
-    name: 'ContractsPage',
-
-    components: {
-        ContractsTable: defineAsyncComponent(
-            () => import('@/components/contracts/Table')
-        )
-    }
-};
-</script>
