--- frontend-vue/src/views/dashboard/AdminPart.vue
@@ -1,17 +1,0 @@
-<template>
-    <users-table />
-</template>
-
-<script>
-import { defineAsyncComponent } from 'vue';
-
-export default {
-    name: 'AdminPart',
-
-    components: {
-        UsersTable: defineAsyncComponent(
-            () => import('@/components/users/Table')
-        )
-    }
-};
-</script>
