+++ frontend-vue/src/views/pages/contracts/TablePage.vue
@@ -4,10 +4,10 @@ import { mapState, mapActions } from 'pinia';
 
 import { useAuthStore } from '@/stores/auth';
 import { useContractStore } from '@/stores/contract';
-import BaseTable from '@/components/common/BaseTable';
+import BaseTable from '@/components/pages/BaseTablePage';
 
 export default {
-    name: 'ContractsTable',
+    name: 'TablePage',
 
     components: {
         AddEditDialog: defineAsyncComponent(
